requirements/base.txt
Flask==1.1.1
python-dotenv==0.10.3
python-decouple==3.3
SQLAlchemy==1.3.12
alembic==1.3.2
Authlib==0.13
Flask-RESTful==0.3.7
peewee==3.13.1
Cerberus==1.3.2
flake8==3.7.9
